Déplace la fonction get_client_ip
