+++ zds/middlewares/matomomiddleware.py
@@ -7,7 +7,7 @@ from _datetime import datetime
 from django.conf import settings
 from django.urls import reverse
 
-from zds.member.views import get_client_ip
+from zds.member.utils import get_client_ip
 from zds.utils.misc import is_ajax
 
 matomo_token_auth = settings.ZDS_APP["site"]["matomo_token_auth"]
